refactor: ScheduleImage를 파일로 분리
