+++ PersonalScheduler/Scene/Namespace/ScheduleImage.swift
@@ -1,0 +1,16 @@
+//
+//  ScheduleImage.swift
+//  PersonalScheduler
+//
+//  Created by Judy on 2023/01/13.
+//
+
+import UIKit
+
+enum ScheduleImage {
+    static let notifyingBell = UIImage(systemName: "bell.circle.fill")
+    static let notifiedBell = UIImage(systemName: "bell.circle")
+    static let unnotifyingBell = UIImage(systemName: "bell.slash.circle.fill")
+    static let add = UIImage(systemName: "plus")
+    static let kakaoButton = UIImage(named: "kakao_login_large_narrow")
+}
