+++ PersonalScheduler/Scene/ScheduleListScene/SubViews/ScheduleListTableViewCell.swift
@@ -180,9 +180,4 @@ extension ScheduleListTableViewCell {
     }
 }
 
-enum ScheduleImage {
-    static let notifyingBell = UIImage(systemName: "bell.circle.fill")
-    static let notifiedBell = UIImage(systemName: "bell.circle")
-    static let unnotifyingBell = UIImage(systemName: "bell.slash.circle.fill")
-    static let add = UIImage(systemName: "plus")
-}
+
